+++ tests/test_pdfutil.py
@@ -5,6 +5,7 @@ import pytest
 
 from barnhunt.pdfutil import (
     concat_pdfs,
+    two_up,
     )
 
 
@@ -38,6 +39,22 @@ def test_concat_pdfs_no_pdfs(tmpdir):
     output_fn = str(tmpdir.join('empty.pdf'))
     with pytest.raises(ValueError):
         concat_pdfs([], output_fn)
+
+
+def test_two_up_two_pages(tmpdir, pdf1, pdf2):
+    out_path = tmpdir.join('output.pdf')
+    in_files = [open(pdf1, 'rb'), open(pdf2, 'rb')]
+    with out_path.open('wb') as out_file:
+        two_up(in_files, out_file)
+    assert page_count(str(out_path)) == 1
+
+
+def test_two_up_three_pages(tmpdir, pdf1, pdf2):
+    out_path = tmpdir.join('output.pdf')
+    in_files = [open(pdf1, 'rb'), open(pdf1, 'rb'), open(pdf2, 'rb')]
+    with out_path.open('wb') as out_file:
+        two_up(in_files, out_file)
+    assert page_count(str(out_path)) == 2
 
 
 def page_count(pdf_fn):